refactor: float patch size

+++ src/include/tlct/convert/helper/record.hpp
@@ -3,7 +3,7 @@
 namespace tlct::_cvt {
 
 struct PsizeMetric {
-    int psize;
+    float psize;
     float metric;
 };
